+++ app/src/main/java/com/example/androidtechsample/MainActivity.kt
@@ -12,86 +12,8 @@ import java.util.concurrent.Executor
 
 class MainActivity : AppCompatActivity() {
 
-    private lateinit var executor: Executor
-    private lateinit var biometricPrompt: BiometricPrompt
-    private lateinit var promptInfo: BiometricPrompt.PromptInfo
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        checkBiometric()
-        createBiometricPrompt()
-        showId()
-    }
-
-    //端末が生体認証に対応しているか確認
-    private fun checkBiometric(): Boolean {
-        val biometricManager = BiometricManager.from(this)
-        when (biometricManager.canAuthenticate()) {
-            BiometricManager.BIOMETRIC_SUCCESS -> {
-                Log.d("TAG", "App can authenticate using biometrics.")
-            }
-            BiometricManager.BIOMETRIC_ERROR_NO_HARDWARE -> {
-                Log.e("TAG", "Hardware not available")
-            }
-            BiometricManager.BIOMETRIC_ERROR_HW_UNAVAILABLE -> {
-                Log.e("TAG", "Biometric features are currently unavailable.")
-            }
-            BiometricManager.BIOMETRIC_ERROR_NONE_ENROLLED -> {
-                Log.e(
-                    "TAG",
-                    "The user hasn't associated any biometric credentials with their account."
-                )
-            }
-            else -> {
-                Log.e("TAG", "Nothing supported")
-            }
-        }
-        return biometricManager.canAuthenticate() == BiometricManager.BIOMETRIC_SUCCESS
-    }
-
-    private fun createBiometricPrompt(): BiometricPrompt {
-        executor = ContextCompat.getMainExecutor(this)
-        val callback = object : BiometricPrompt.AuthenticationCallback() {
-            override fun onAuthenticationError(errorCode: Int, errString: CharSequence) {
-                super.onAuthenticationError(errorCode, errString)
-                Toast.makeText(
-                    applicationContext,
-                    "Authentication error: $errString", Toast.LENGTH_SHORT
-                ).show()
-            }
-
-            override fun onAuthenticationFailed() {
-                super.onAuthenticationFailed()
-                Toast.makeText(
-                    applicationContext, "Authentication failed",
-                    Toast.LENGTH_SHORT
-                ).show()
-            }
-
-            override fun onAuthenticationSucceeded(result: BiometricPrompt.AuthenticationResult) {
-                super.onAuthenticationSucceeded(result)
-                Toast.makeText(
-                    applicationContext,
-                    "Authentication succeeded!", Toast.LENGTH_SHORT
-                ).show()
-            }
-        }
-        biometricPrompt = BiometricPrompt(this, executor, callback)
-        return biometricPrompt
-    }
-
-    private fun showId() {
-        promptInfo = BiometricPrompt.PromptInfo.Builder()
-            .setTitle("指紋認証")
-            .setSubtitle("サンプルです")
-            .setConfirmationRequired(true)
-            .setDeviceCredentialAllowed(true)
-            .build()
-        val biometricLoginButton =
-            findViewById<Button>(R.id.button_hand)
-        biometricLoginButton.setOnClickListener {
-            biometricPrompt.authenticate(promptInfo)
-        }
     }
 }
